rescheduling hook up ui display

+++ api/src/apis/group_user_schedules.ts
@@ -1,4 +1,4 @@
-import { asyncForEach, IGroupUserScheduleActionTypes, IGroupService, IGroupUserSchedule } from 'awayto';
+import { asyncForEach, IGroupUserScheduleActionTypes, IGroupService, IGroupUserSchedule, ScheduledParts, IGroupUserScheduleStub, IGroupUserScheduleState } from 'awayto';
 import { ApiModule } from '../api';
 
 const groupUserSchedules: ApiModule = [
@@ -61,6 +61,30 @@ const groupUserSchedules: ApiModule = [
   },
 
   {
+    action: IGroupUserScheduleActionTypes.GET_GROUP_USER_SCHEDULE_STUBS,
+    cache: 'skip',
+    cmnd: async (props) => {
+      try {
+        const { groupName } = props.event.pathParameters;
+
+        const response = await props.db.query<IGroupUserScheduleStub>(`
+          SELECT guss.*
+          FROM dbview_schema.group_user_schedule_stubs guss
+          JOIN dbtable_schema.schedules schedule ON schedule.id = guss."groupScheduleId"
+          JOIN dbview_schema.enabled_users eu ON eu.sub = schedule.created_sub
+          WHERE eu.username = $1
+        `, ['system_group_' + groupName]);
+
+        return { stubs: response.rows } as IGroupUserScheduleState;
+
+      } catch (error) {
+        throw error;
+      }
+
+    }
+  },
+
+  {
     action: IGroupUserScheduleActionTypes.DELETE_GROUP_USER_SCHEDULE_BY_USER_SCHEDULE_ID,
     cmnd: async (props) => {
       try {
@@ -69,12 +93,28 @@ const groupUserSchedules: ApiModule = [
         const idsSplit = ids.split(',');
 
         await asyncForEach(idsSplit, async userScheduleId => {
-          // Detach user schedule from group
-          await props.db.query<IGroupService>(`
-            DELETE FROM dbtable_schema.group_user_schedules
-            WHERE user_schedule_id = $1
-            RETURNING id
+
+          const { rows: parts } = await props.db.query<ScheduledParts>(`
+            SELECT * FROM dbfunc_schema.get_scheduled_parts($1);
           `, [userScheduleId]);
+
+          const hasParts = parts.some(p => p.ids?.length);
+
+          if (!hasParts) {
+            // Detach user schedule from group
+            await props.db.query<IGroupService>(`
+              DELETE FROM dbtable_schema.group_user_schedules
+              WHERE user_schedule_id = $1
+              RETURNING id
+            `, [userScheduleId]);
+          } else {
+            await props.db.query<IGroupService>(`
+              UPDATE dbtable_schema.group_user_schedules
+              SET enabled = false
+              WHERE user_schedule_id = $1
+              RETURNING id
+            `, [userScheduleId]);
+          }
         });
 
         await props.redis.del(props.event.userSub + `group/${groupName}/schedules`);
